!1082 [自研][PyTorch]Faster_Mask_RCNN_for_PyTorch脚本修复 !1082 [自研][PyTorch]Faster_Mask_RCNN_for_PyTorch脚本修复

+++ PyTorch/built-in/cv/detection/Faster_Mask_RCNN_for_PyTorch/detectron2/structures/boxes.py
@@ -316,7 +316,7 @@ class Boxes:
         assert all([isinstance(box, Boxes) for box in boxes_list])
 
         # use torch.cat (v.s. layers.cat) so the returned boxes never share storage with input
-        cat_boxes = cls(torch.cat([b.tensor for b in boxes_list], dim=0))
+        cat_boxes = cls(torch.cat([b.tensor.float() for b in boxes_list], dim=0))
         return cat_boxes
 
     @property
